Implement BLS disabled flag for the SpecConfig
This flag allows to run spec tests with BLS disabled flag.
The flat is particularly utilized in the following places:
* AbstractBlockProcessor.processAndValidateBlock when processing a block
* ForkChoiceUtil.validate when validating an attestation
* BeaconStateAccessorsAltair.getNextSyncCommitteeIndices when computing a sync committee aggregate pubkey

The decision to put it this way is mainly because of a lot of boilerplate code that would be required to
propagate signature verifier to the downstream method calls. Instead of propagating it is read from
the spec config instance available at hand.

TestDefinition exposes an interface to create a spec with blsDisabled flag.